Add copy_to_www.sh
copy_to_www.sh copies all the project files to the /home/www/data folder
on dwtj.me. It performs a local copy if you're on dwtj.me, and uses scp
if you're not.

This script will need to be updated when new files are added to the
project, and will not work unless you're somewhere inside the git
repository.

Also, you need to alias dwtj.me (with the appropriate username) to
"dwtj" in your .ssh/config file if you're using this script to push from
your computer.