Change layout so terminal and GUI each take 50%
Change the GUI layout so that the terminal and the GUI each occupy 50%
of the vertical screen space. Also, both now occupy the full width of
the window.

Notes:
- I used CSS flexboxes because it's not possible to sanely specify
layout in CSS without them. This unfortunately ties our code to Chrome
(and possibly Safari?). To those editing this code: the best resource
on flexboxes I know of is the spec itself, which is at http://www.w3.org/TR/css3-
flexbox/. Note that there are multiple revisions of the spec; see
http://css-tricks.com/old-flexbox-and-new-flexbox/