Carriage return -> line feed translation
The ShellInABox code on the client sends carriage returns when [Enter]
is pressed, but the backend expects line feeds (new lines), so
translate.

Also, whenever the backend sends a line feed, it needs to be translated
into a carriage return plus a line feed before being sent to
ShellInABox's code.